Moved common interop code, adding missing html/refs
